gpt4all-backend/requirements.txt
fastapi
uvicorn
gpt4all==1.0.1
google-cloud-storage 